// server/@types/shared/models/Adjudication.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Adjudication = {
    establishment: string;
    finding?: string | null;
    hearingHeld: boolean;
    id: number;
    offenceDescription: string;
    reportedAt: string;
};